requirements.txt
certifi==2018.10.15
Django==2.1.4
django-ranged-response==0.2.0
django-simple-captcha==0.5.9
Pillow==5.3.0
pytz==2018.7
six==1.12.0
wincertstore==0.2
